Refactor tests to match new structure
