+++ tests/domain/test_model.py
@@ -3,196 +3,9 @@ import pytest
 from dbt_airflow.domain.model import (
     AirflowTask,
     DbtAirflowTask,
-    DbtResourceType,
-    Manifest,
-    NodeDeps,
     TaskList,
 )
-
-
-@pytest.fixture
-def mock_manifest_data():
-    return {
-        'metadata': {},
-        'nodes': {
-            'model.mypackage.my_model': {
-                'name': 'my_model',
-                'resource_type': 'model',
-                'compiled': True,
-                'depends_on': {
-                    'macros': [],
-                    'nodes': [
-                        'seed.mypackage.my_seed',
-                        'model.mypackage.another_model'
-                    ]
-                },
-                'package_name': 'mypackage',
-                'fqn': [
-                    'a',
-                    'b',
-                    'c'
-                ]
-            },
-            'seed.mypackage.my_seed': {
-                'name': 'my_seed',
-                'resource_type': 'seed',
-                'compiled': True,
-                'depends_on': {
-                    'macros': [],
-                    'nodes': []
-                },
-                'package_name': 'mypackage',
-                'fqn': [
-                    'a'
-                ]
-            },
-            'model.mypackage.another_model': {
-                'name': 'another_model',
-                'resource_type': 'model',
-                'compiled': True,
-                'depends_on': {
-                    'macros': [],
-                    'nodes': [
-                        'seed.mypackage.my_seed'
-                    ]
-                },
-                'package_name': 'mypackage',
-                'fqn': ['d', 'e', 'f']
-            },
-            'snapshot.mypackage.my_snapshot': {
-                'name': 'my_snapshot',
-                'resource_type': 'snapshot',
-                'compiled': True,
-                'depends_on': {
-                    'macros': [],
-                    'nodes': [
-                        'model.mypackage.my_model'
-                    ]
-                },
-                'package_name': 'mypackage',
-                'fqn': ['a', 'b', 'c']
-            },
-            'test.mypackage.not_null_another_model_field_A.c9c3c572df': {
-                'name': 'not_null_another_model_field_A',
-                'resource_type': 'test',
-                'compiled': True,
-                'depends_on': {
-                    'macros': [],
-                    'nodes': [
-                        'model.mypackage.another_model'
-                    ]
-                },
-                'package_name': 'mypackage',
-                'fqn': ['w', 'x', 'y']
-            }
-        }
-    }
-
-
-@pytest.fixture
-def mock_manifest(mock_manifest_data):
-    return Manifest(**mock_manifest_data)
-
-
-@pytest.fixture
-def mock_node(mock_manifest):
-    return mock_manifest.nodes['model.mypackage.my_model']
-
-
-@pytest.fixture
-def mock_node_deps(mock_node):
-    return mock_node.depends_on
-
-
-@pytest.fixture
-def mock_dbt_airflow_task():
-
-    def create_task(
-        task_id='model.mypackage.my_model',
-        manifest_node_name='model.mypackage.my_model',
-        resource_type=DbtResourceType.model,
-        upstream_task_ids=None,
-        task_group='b',
-        package_name='mypackage',
-    ):
-        if not upstream_task_ids:
-            upstream_task_ids = {'seed.mypackage.my_seed', 'model.mypackage.another_model'}
-
-        return DbtAirflowTask(
-            task_id=task_id,
-            manifest_node_name=manifest_node_name,
-            resource_type=resource_type,
-            upstream_task_ids=upstream_task_ids,
-            task_group=task_group,
-            package_name=package_name,
-        )
-
-    return create_task
-
-
-def test_node_deps_model_filters_keeps_only_model_seed_and_snapshot_resources():
-    """
-    GIVEN a list of nodes representing NodeDeps
-    WHEN validating `nodes` field of NodeDeps dataclass
-    THEN the validator will filter out any dependency which does not correspond to a dbt model,
-        seed or snapshot
-    """
-    # GIVEN
-    resources = [
-        'model.profile_name.my_model',
-        'model.profile_name.another_model',
-        'seed.profile_name.my_seed',
-        'snapshot.profile_name.my_snapshot',
-        'test.profile_name.not_null_test_my_model',
-    ]
-
-    # WHEN
-    node_deps = NodeDeps(nodes=resources)
-
-    # THEN
-    assert node_deps.nodes == [
-        'model.profile_name.my_model',
-        'model.profile_name.another_model',
-        'seed.profile_name.my_seed',
-        'snapshot.profile_name.my_snapshot',
-    ]
-
-
-def test_node_model_parsing(mock_node):
-    assert mock_node.name == 'my_model'
-    assert mock_node.resource_type == 'model'
-    assert mock_node.package_name == 'mypackage'
-    assert mock_node.task_group == 'b'
-    assert len(mock_node.depends_on.nodes) == 2
-
-
-def test_manifest_model_parsing(mock_manifest_data):
-    manifest = Manifest(**mock_manifest_data)
-    assert len(manifest.nodes) == 5
-
-
-def test_manifest_model_get_statistics(mock_manifest):
-    """
-    GIVEN manifest json data
-    WHEN parsing the Manifest model
-    THEN the number of parsed resource types (models, seeds, snapshot, tests) is computed correctly
-    """
-    # GIVEN/WHEN
-    actual = mock_manifest.get_statistics()
-
-    # THEN
-    expected = {
-        'models': 2,
-        'tests': 1,
-        'snapshots': 1,
-        'seeds': 1,
-    }
-    assert actual == expected
-
-
-def test_node_deps_model_parsing(mock_node_deps):
-    assert len(mock_node_deps.nodes) == 2
-    assert mock_node_deps.nodes == ['seed.mypackage.my_seed', 'model.mypackage.another_model']
+from dbt_airflow.parser.dbt import DbtResourceType
 
 
 def test_airflow_task_equality():
